refactor side drawer into smaller components, add link to import
