+++ locales/en.json
@@ -20,6 +20,7 @@
   "thisLanguage": "english",
   "switchTheme": "switch theme",
   "upload": {
+    "word": "upload",
     "inProgress": "File is uploading. Do not worry, this may take a couple of seconds.",
     "done": "Your token is {0}. You may share it with others to give them access to your Slack archive.",
     "button": "upload export (.zip)"
